download excel file

+++ server.js
@@ -2,6 +2,8 @@ const express = require('express');
 var bodyParser = require('body-parser');
 const fs = require('fs');
 const path = require('path');
+const stream = require('stream');
+const xlsx = require('node-xlsx').default;
 
 const app = express();
 
@@ -10,9 +12,7 @@ const fileName = 'db.json';
 app.use((req, res, next) => {
   var now = new Date().toString();
   var log = `${now}: ${req.method} ${req.url}`;
-
   console.log(log);
-  fs.appendFile('server.log', log + '\n',(error) => {});
   next();
 });
 
@@ -30,20 +30,60 @@ app.get('/', (req, res) => {
 
 app.post('/save', (req, res) => {
   fs.exists(fileName, (exists) => {
-    fs.readFile(fileName, (err, data) => {
-      let allData;
-      if (data) {
-        allData = JSON.parse(data);
-      } else {
-        allData = {data:[]};
-      }
-      allData.data.push(req.body);
-      fs.writeFile(fileName, JSON.stringify(allData),(error) => {});
-    })
+    if (exists) {
+      fs.readFile(fileName, (err, data) => {
+        let allData;
+        if (data) {
+          allData = JSON.parse(data);
+        } else {
+          allData = {
+            data: []
+          };
+        }
+        allData.data.push(req.body);
+        fs.writeFile(fileName, JSON.stringify(allData), (error) => {});
+      });
+    }
   });
   res.json({
     msg: 'done'
   });
+});
+
+app.get('/download', (req, res) => {
+  fs.exists(fileName, (exists) => {
+    if (exists) {
+      fs.readFile(fileName, (err, data) => {
+        let allData;
+        if (data) {
+          allData = JSON.parse(data)['data'];
+        } else {
+          allData = [];
+        }
+        const excelRows = allData.map((dbRow) => {
+          return Object.values(dbRow);
+        });
+        const buffer = xlsx.build([{
+          name: 'Saved Data',
+          data: excelRows
+        }]);
+        var fileContents = Buffer.from(buffer, "base64");
+        const readStream = new stream.PassThrough();
+        readStream.end(fileContents);
+
+        res.set('Content-disposition', 'attachment; filename=SavedData.xlsx');
+        res.set('Content-Type', 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
+
+        readStream.pipe(res);
+      });
+    } else {
+      res.send('nothing to show');
+    }
+  });
+});
+
+app.get('/reset-db', (req, res) => {
+  fs.writeFile(fileName, JSON.stringify({}), (error) => {});
 });
 
 app.listen(3000, () => {
